refactor: moving content of employers/[employeeID] to add/page.tsx
